// Presentation/cosmeticshop.client/src/apiClient/models/product-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CosmeticShop.WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ProductRequestDto
 */
export interface ProductRequestDto {

    /**
     * @type {string}
     * @memberof ProductRequestDto
     */
    name: string;

    /**
     * @type {string}
     * @memberof ProductRequestDto
     */
    description: string;

    /**
     * @type {number}
     * @memberof ProductRequestDto
     */
    price?: number;

    /**
     * @type {string}
     * @memberof ProductRequestDto
     */
    categoryId: string;

    /**
     * @type {number}
     * @memberof ProductRequestDto
     */
    stockQuantity?: number;

    /**
     * @type {number}
     * @memberof ProductRequestDto
     */
    rating?: number;

    /**
     * @type {string}
     * @memberof ProductRequestDto
     */
    manufacturer?: string | null;

    /**
     * @type {string}
     * @memberof ProductRequestDto
     */
    imageUrl?: string | null;
}