// Presentation/cosmeticshop.client/src/apiClient/models/product-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CosmeticShop.WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ProductResponseDto
 */
export interface ProductResponseDto {

    /**
     * @type {string}
     * @memberof ProductResponseDto
     */
    id: string;

    /**
     * @type {string}
     * @memberof ProductResponseDto
     */
    name: string;

    /**
     * @type {string}
     * @memberof ProductResponseDto
     */
    description: string;

    /**
     * @type {number}
     * @memberof ProductResponseDto
     */
    price: number;

    /**
     * @type {string}
     * @memberof ProductResponseDto
     */
    categoryId: string;

    /**
     * @type {number}
     * @memberof ProductResponseDto
     */
    stockQuantity: number;

    /**
     * @type {number}
     * @memberof ProductResponseDto
     */
    rating: number;

    /**
     * @type {string}
     * @memberof ProductResponseDto
     */
    manufacturer: string;

    /**
     * @type {string}
     * @memberof ProductResponseDto
     */
    imageUrl: string;
}